Rename InstanceRepository => ExecutorRepository, add jsdocs
