+++ cloud/wfm/src/executor/Executor.ts
@@ -7,8 +7,18 @@ import Process from '../process/Process';
  * Triggers instantiation and execution of a {@link ProcessInstance}
  */
 interface Executor {
+  /**
+   * {@link Process} to be executed
+   */
   process: Process;
+  /**
+   * {@link ProcessInstance} that is created based on {@link Process}
+   */
   instance?: ProcessInstance;
+
+  /**
+   * Starts the execution of the assignend Process
+   */
   start(): void;
 }
