+++ cloud/wfm/src/executor/ExecutorImpl.ts
@@ -3,10 +3,17 @@ import ProcessInstance from '../process-instance/ProcessInstance';
 import Process from '../process/Process';
 import Task, {TaskEventData, TaskStatus} from '../task/Task';
 import Executor from './Executor';
-import ExecutorRepository from './InstanceRepository';
+import ExecutorRepository from './ExecutorRepository';
 
+/**
+ * Default implementation for {@link Executor}
+ */
 export default class ExecutorImpl implements Executor {
   public instance: ProcessInstance;
+  /**
+   * @param process The {@link Process} this {@link Executor} is to run
+   * @param repository {@link ExecutorRepository} instance for data access
+   */
   constructor(public process: Process, public repository: ExecutorRepository) {
     this.instance = this.process.createInstance();
   }
@@ -20,17 +27,31 @@ export default class ExecutorImpl implements Executor {
         .then(() => this.runCurrentTask());
     }
   }
+
+  /**
+   * Orders {@link repository} to update {@link instance}
+   */
   protected saveInstance() {
     return this.repository.saveInstance(this.instance);
   }
+
+  /**
+   * Runs the present {@link Task} of {@link instance} and sets up execution of
+   * the next one as it is run
+   */
   protected runCurrentTask() {
     this.instance.getCurrentTask()
-    .then(t => {
+    .then(task => {
       // currentTask is undefined after all are done
-      if (!t) {
+      if (!task) {
         return;
       }
-      return t.on('statusChange', this.onTaskDone.bind(this)).run();
+      return task
+        // TODO: this assumes the Task object will be kept in memory forever
+        // but human-executed tasks can take very long to complete
+        // we might have to change this to an out-of-process message bus
+        .on('statusChange', this.onTaskDone.bind(this))
+        .run();
     });
   }
 }
